+++ ui/src/app/desktop-document-mode.ts
@@ -1,12 +1,10 @@
 import { normalizeRouteBasePath, normalizeRoutePath } from "@openclaw/uirouter";
-import type { GatewaySessionRow } from "../api/types.ts";
-import { resolveChatPaneDesktopTarget } from "../pages/chat/chat-pane-placement.ts";
 
 const DESKTOP_DOCUMENT_PATH = "/desktop";
 
 type DesktopDocumentLocation = Pick<Location, "pathname" | "search">;
 
-type DesktopDocumentOptions = {
+export type DesktopDocumentOptions = {
   source: string | null;
   session: string | null;
   control: boolean;
@@ -39,11 +37,4 @@ export function desktopDocumentOptions(
     session: search.get("session"),
     control: search.get("control") === "1",
   };
-}
-
-export function resolveDesktopDocumentTarget(
-  options: DesktopDocumentOptions,
-  session: GatewaySessionRow | undefined,
-): string | null {
-  return options.source ?? (options.session ? resolveChatPaneDesktopTarget(session) : null);
 }
